Added Page folder with Homepage and new creation page so not everything is in the Main.Dart file.
not finished look wise but i want to keep it very minimalistic and clean in design, I want to add a dark mode switch as well for dark theme option.
