+++ lib/Pages/create_pages.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class CreatePage extends StatelessWidget {
+  const CreatePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(centerTitle: true, title: Text('Create new task')),
+    );
+  }
+}
