+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:template/Pages/home_page.dart';
 
 void main() {
   runApp(const MyApp());
@@ -7,98 +8,11 @@ void main() {
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        colorScheme:
-            ColorScheme.fromSeed(seedColor: const Color.fromARGB(255, 0, 0, 0)),
-        useMaterial3: true,
-      ),
-      home: const MyHomePage(title: 'TIG333 TODO - Mattias'),
+    return const MaterialApp(
+      debugShowCheckedModeBanner: false,
+      home: HomePage(),
     );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key, required this.title});
-
-  // This widget is the home page of your application. It is stateful, meaning
-  // that it has a State object (defined below) that contains fields that affect
-  // how it looks.
-
-  // This class is the configuration for the state. It holds the values (in this
-  // case the title) provided by the parent (in this case the App widget) and
-  // used by the build method of the State. Fields in a Widget subclass are
-  // always marked "final".
-
-  final String title;
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
-
-  void _incrementCounter() {
-    setState(() {
-      _counter++;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-          backgroundColor: Colors.black,
-          title: Text(
-            widget.title,
-            style: TextStyle(color: Colors.white),
-          ),
-          centerTitle: true,
-        ),
-        body: ListView.builder(
-          itemCount: 5,
-          itemBuilder: (context, index) {
-            return Padding(
-              padding: const EdgeInsets.all(10.0),
-              child: Row(
-                children: [
-                  Padding(
-                    //Vill tillägga dynamisk padding med MediaQuery classen men inte säker hur man gör än :P
-                    padding: const EdgeInsets.only(right: 10.0),
-                    child: Checkbox(
-                      value: false,
-                      onChanged: (bool? value) {},
-                    ),
-                  ),
-                  Expanded(child: Text('Task $index')),
-                  IconButton(
-                    icon: Icon(Icons.delete),
-                    onPressed: () {},
-                  ),
-                ],
-              ),
-            );
-          },
-        ),
-        floatingActionButton: Padding(
-          padding: const EdgeInsets.only(right: 10.0, bottom: 40),
-          child: SizedBox(
-            width: 70.0,
-            height: 70.0,
-            child: FloatingActionButton(
-              onPressed: () {
-                //Lägg till action när nedtryckt
-              },
-              backgroundColor: Colors.black,
-              foregroundColor: Colors.white,
-              child: Icon(Icons.add),
-            ),
-          ),
-        ));
   }
 }
